Split the list screen up into different files
The list screen will show different panes based on its current state, update to split into multiple
files to more easily model this.
