+++ feature/discover/src/commonMain/kotlin/com/chesire/nekomp/feature/discover/ui/DiscoverScreen.kt
@@ -7,30 +7,19 @@
 package com.chesire.nekomp.feature.discover.ui
 
 import androidx.compose.animation.AnimatedContent
+import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionLayout
-import androidx.compose.foundation.horizontalScroll
+import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.Card
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ElevatedButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.OutlinedTextField
@@ -48,15 +37,16 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.backhandler.BackHandler
-import androidx.compose.ui.graphics.vector.rememberVectorPainter
-import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.focus.onFocusEvent
+import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
 import kotlinx.collections.immutable.ImmutableList
 import org.koin.compose.viewmodel.koinViewModel
 
@@ -69,18 +59,6 @@ fun DiscoverScreen(viewModel: DiscoverViewModel = koinViewModel()) {
         execute = { viewModel.execute(it) }
     )
 }
-
-// List/Detail/Support(?)
-// List shows a search bar at the top, with the trending series below it
-// Clicking the search bar shows a drop down of search options
-// Executing search goes to the detail screen
-// Detail screen shows the list of searched series
-// Support screen shows the details about a series (?) depending on implementation
-
-// Default list screen should show the current trending series
-// Search bar at top of list, clicking it should change to a search screen
-// Add a top rated section?
-// Clicking on a trending item shows the details for it?
 
 @Composable
 private fun Render(
@@ -141,7 +119,7 @@ private fun Render(
                     },
                     detailPane = {
                         AnimatedPane {
-                            DetailContent(
+                            DetailPane(
                                 item = navigator.currentDestination?.content,
                                 showBack = navigator.scaffoldValue.primary == PaneAdaptedValue.Expanded,
                                 goBack = { navigator.navigateBack() }
@@ -152,6 +130,12 @@ private fun Render(
             }
         }
     }
+}
+
+private enum class ListPaneType {
+    Trending,
+    Search,
+    Results
 }
 
 @Composable
@@ -165,163 +149,74 @@ private fun ListContent(
     onItemClick: (DiscoverItem) -> Unit,
     onTrackClick: (DiscoverItem) -> Unit
 ) {
+    val focus = LocalFocusManager.current
+    var displayedPaneType by remember { mutableStateOf(ListPaneType.Trending) }
+    BackHandler(enabled = displayedPaneType != ListPaneType.Trending) {
+        when (displayedPaneType) {
+            ListPaneType.Trending -> Unit
+            ListPaneType.Search -> {
+                displayedPaneType = ListPaneType.Trending
+                focus.clearFocus(true)
+            }
+
+            ListPaneType.Results -> {
+                displayedPaneType = ListPaneType.Search
+            }
+        }
+    }
+
     Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-        OutlinedTextField(
-            value = "",
-            onValueChange = {},
-            modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp),
-            label = {
-                Text("Search")
-            },
-            leadingIcon = {
-                Icon(imageVector = Icons.Default.Search, contentDescription = null)
-            }
-        )
-        Column(
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-            modifier = Modifier.verticalScroll(rememberScrollState())
-        ) {
-            TrendingSection(
-                title = "Trending anime",
-                items = trendingAnime,
-                onItemClick = onItemClick,
-                onTrackClick = onTrackClick
-            )
-            TrendingSection(
-                title = "Trending manga",
-                items = trendingManga,
-                onItemClick = onItemClick,
-                onTrackClick = onTrackClick
-            )
-            TrendingSection(
-                title = "Top rated anime",
-                items = topRatedAnime,
-                onItemClick = onItemClick,
-                onTrackClick = onTrackClick
-            )
-            TrendingSection(
-                title = "Top rated manga",
-                items = topRatedManga,
-                onItemClick = onItemClick,
-                onTrackClick = onTrackClick
-            )
-            TrendingSection(
-                title = "Most popular anime",
-                items = mostPopularAnime,
-                onItemClick = onItemClick,
-                onTrackClick = onTrackClick
-            )
-            TrendingSection(
-                title = "Most popular manga",
-                items = mostPopularManga,
-                onItemClick = onItemClick,
-                onTrackClick = onTrackClick
-            )
-        }
-    }
-}
-
-@Composable
-private fun TrendingSection(
-    title: String,
-    items: ImmutableList<DiscoverItem>,
-    onItemClick: (DiscoverItem) -> Unit,
-    onTrackClick: (DiscoverItem) -> Unit
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            text = title,
-            modifier = Modifier.padding(horizontal = 16.dp)
-        )
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(IntrinsicSize.Max)
-                .horizontalScroll(rememberScrollState())
-                .safeContentPadding(),
-            horizontalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            items.forEach {
-                TrendingDisplay(
-                    discoverItem = it,
-                    modifier = Modifier.fillMaxHeight(),
-                    onItemClick = onItemClick,
-                    onTrackClick = onTrackClick
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun TrendingDisplay(
-    discoverItem: DiscoverItem,
-    modifier: Modifier = Modifier,
-    onItemClick: (DiscoverItem) -> Unit,
-    onTrackClick: (DiscoverItem) -> Unit
-) {
-    Card(
-        onClick = { onItemClick(discoverItem) },
-        modifier = modifier.width(256.dp)
-    ) {
-        Box {
-            AsyncImage(
-                model = discoverItem.coverImage,
-                contentDescription = null,
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.FillHeight,
-                alpha = 0.3f
-            )
-            Column(modifier = Modifier.fillMaxSize().padding(8.dp)) {
-                // image
-                Text(discoverItem.title) // title
-                // current rating
-                // Synopsis
-                // track button
-                Spacer(Modifier.weight(1f))
-                if (!discoverItem.isTracked) {
-                    ElevatedButton(
-                        onClick = { onTrackClick(discoverItem) },
-                        modifier = Modifier.align(Alignment.End),
-                        enabled = !discoverItem.isPendingTrack
-                    ) {
-                        if (discoverItem.isPendingTrack) {
-                            CircularProgressIndicator()
-                        } else {
-                            Text("Track")
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun DetailContent(
-    item: DiscoverItem?,
-    showBack: Boolean,
-    goBack: () -> Unit
-) {
-    Column(
-        modifier = Modifier.padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        if (item != null) {
-            if (showBack) {
+        Row {
+            AnimatedVisibility(
+                visible = displayedPaneType != ListPaneType.Trending,
+                modifier = Modifier.align(Alignment.CenterVertically)
+            ) {
                 IconButton(
-                    onClick = goBack,
-                    modifier = Modifier.align(Alignment.Start)
+                    onClick = {
+                        displayedPaneType = ListPaneType.Trending
+                        focus.clearFocus(true)
+                    }
                 ) {
                     Icon(
-                        painter = rememberVectorPainter(Icons.AutoMirrored.Filled.ArrowBack),
-                        contentDescription = "Go back"
+                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                        contentDescription = null
                     )
                 }
             }
-        } else {
-            Text("No entry selected")
+            OutlinedTextField(
+                value = "",
+                onValueChange = {},
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp)
+                    .animateContentSize()
+                    .onFocusEvent {
+                        if (it.hasFocus) {
+                            displayedPaneType = ListPaneType.Search
+                        }
+                    },
+                label = {
+                    Text("Search")
+                },
+                leadingIcon = {
+                    Icon(imageVector = Icons.Default.Search, contentDescription = null)
+                }
+            )
+        }
+        when (displayedPaneType) {
+            ListPaneType.Trending -> TrendingPane(
+                trendingAnime = trendingAnime,
+                trendingManga = trendingManga,
+                topRatedAnime = topRatedAnime,
+                topRatedManga = topRatedManga,
+                mostPopularAnime = mostPopularAnime,
+                mostPopularManga = mostPopularManga,
+                onItemClick = onItemClick,
+                onTrackClick = onTrackClick
+            )
+
+            ListPaneType.Search -> SearchPane()
+            ListPaneType.Results -> ResultsPane()
         }
     }
 }
