+++ feature/discover/src/commonMain/kotlin/com/chesire/nekomp/feature/discover/ui/ResultsPane.kt
@@ -1,0 +1,8 @@
+package com.chesire.nekomp.feature.discover.ui
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun ResultsPane() {
+    
+}
