+++ feature/discover/src/commonMain/kotlin/com/chesire/nekomp/feature/discover/ui/SearchPane.kt
@@ -1,0 +1,15 @@
+package com.chesire.nekomp.feature.discover.ui
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun SearchPane() {
+    Column {
+        Text("Test1")
+        Text("Test2")
+        Text("Test3")
+        Text("Test4")
+    }
+}
